+++ examples/4. recover-from-stop/index.ts
@@ -1,29 +1,35 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
+import _ = require('lodash');
 import Spider from '../../src/core/Spider';
 import logger from '../../src/util/logger';
-// @ts-ignore
-import TurndownService = require('turndown');
-// @ts-ignore
-import turndownPluginGfm = require('turndown-plugin-gfm');
 
-const turndownService = new TurndownService();
-turndownService.use(turndownPluginGfm.gfm);
-
-new Spider({
+let spider = new Spider({
+    rateLimit: {
+        'nodejs.org': [2, 10], // 2 requests at most in 10 seconds.
+    },
     handlers: [{
-        pattern: 'https://github.com/*/*',
+        pattern: 'https://nodejs.org/api/',
         handle: async function (response, items) {
-            let page = await response.xpath('//article') as string;
-            this.pipe(items.pack({ markdown: page }));
+            let links = await response.xpath('//*[@id="column2"]/ul[2]//a/@href');
+            this.enqueue(..._.map(links, link => response.url + link));
+        },
+    }, {
+        pattern: 'https://nodejs.org/api/*',
+        handle: async function (response, items) {
+            items.html = await response.cssModel({
+                title: 'title::text()',
+                body: '#column1>div::html()'
+            });
+            this.pipe(items);
         },
     }],
     pipelines: [
         items => {
-            if (!items.markdown) return;
+            if (!items.html) return;
 
-            items.file = turndownService.turndown(items.markdown);
-            items.extension = 'md';
+            items.file = `<!DOCTYPE html>\n<html lang="en">\n<head>\n<title>${items.html.title}</title>\n</head>\n<body>\n${items.html.body}\n</body>\n</html>`;
+            items.extension = 'html';
         },
         items => {
             if (!items.file) return;
@@ -45,5 +51,8 @@ new Spider({
         },
     ],
 })
-    .enqueue('https://github.com/plylrnsdy/seeker')
-    .start()
+    .enqueue('https://nodejs.org/api/')
+    .start();
+
+
+setTimeout(() => spider.stop(), 18 * 1000);